+++ servo/cli.py
@@ -1,5 +1,6 @@
 import asyncio
 import json
+import re
 import shlex
 import subprocess
 import sys
@@ -8,7 +9,7 @@ from datetime import datetime, timezone
 from enum import Enum
 from functools import reduce
 from pathlib import Path
-from typing import Any, Awaitable, Callable, Iterable, List, Optional, Set, Type, Tuple, Union
+from typing import Any, Awaitable, Callable, Iterable, List, Optional, Pattern, Set, Type, Tuple, Union
 
 import bullet
 import click
@@ -31,15 +32,16 @@ from servo.assembly import (
 )
 from servo.connector import (
     BaseConnector, 
-    Optimizer,
-    _connector_class_from_string
+    Optimizer
 )
 from servo.events import EventHandler, EventResult, Preposition
 from servo.logging import logger, set_level as set_log_level
 from servo.servo import (
     Events,
     Servo,
+    _connector_class_from_string
 )
+from servo.checks import Check, Filter, HaltOnFailed
 from servo.runner import Runner
 from servo.types import *
 from servo.utilities import PreservedScalarString, commandify
@@ -72,6 +74,7 @@ class LogLevel(str, Enum):
     ERROR = "ERROR"
     CRITICAL = "CRITICAL"
 
+# FIXME: Eliminate the mixin and put our context object onto the Click.obj instance
 class Context(typer.Context):
     """
     Context models state required by different CLI invocations.
@@ -972,6 +975,18 @@ class ServoCLI(CLI):
                 help="Connectors to check",
                 callback=self.connectors_instance_callback,
             ),
+            name: Optional[List[str]] = typer.Option(
+                False, "--name", "-n", help="Filter by name"
+            ),
+            id: Optional[List[str]] = typer.Option(
+                False, "--id", "-i", help="Filter by ID"
+            ),
+            tag: Optional[List[str]] = typer.Option(
+                False, "--tag", "-t", help="Filter by tag"
+            ),
+            halt_on: HaltOnFailed = typer.Option(
+                HaltOnFailed.requirement, "--halt-on-failed", "-h", help="Halt running checks on a failure condition",
+            ),
             verbose: bool = typer.Option(
                 False, "--verbose", "-v", help="Display verbose output"
             ),
@@ -985,8 +1000,6 @@ class ServoCLI(CLI):
             """
             Check that the servo is ready to run
             """
-            # TODO: Requires a config file
-
             # FIXME: temporary workaround until I can unwind Context overload
             if isinstance(context, click.core.Context):
                 context = context.parent
@@ -996,33 +1009,73 @@ class ServoCLI(CLI):
                 validate_connectors_respond_to_event(connectors, Events.CHECK)
             else:
                 connectors = context.assembly.connectors
+            
+            def parse_re(value: Optional[List[str]]) -> Union[None, List[str], Pattern[str]]:
+                if value and len(value) == 1:
+                    val = value[0]
+                    if val[:1] == '/' and val[-1] == '/':
+                        return re.compile(val[1:-1])
+                
+                return value
+            
+            def parse_csv(value: Optional[List[str]]) -> Union[None, List[str], Pattern[str]]:
+                if value and len(value) == 1:
+                    val = value[0]
+                    if "," in val:
+                        return list(map(lambda v: v.strip(), val.split(",")))
+                
+                return value
+            
+            def parse_id(value: Optional[List[str]]) -> Union[None, List[str], Pattern[str]]:
+                v = parse_re(value)
+                if not isinstance(v, Pattern):
+                    return parse_csv(v)
+                
+                return v
 
+            args = dict(name=parse_re(name), id=parse_id(id), tags=parse_csv(tag))
+            constraints = dict(filter(lambda i: bool(i[1]), args.items()))
+            filter_ = Filter(**constraints)
             results: List[EventResult] = sync(context.servo.dispatch_event(
-                Events.CHECK, include=connectors
+                Events.CHECK, filter_, include=connectors, halt_on=halt_on
             ))
 
             table = []
             ready = True
             if verbose:
-                headers = ["CONNECTOR", "CHECK", "STATUS", "COMMENT"]                    
-                for result in results:
-                    checks: List[Check] = result.value                
-                    names, statuses, comments = [], [], []
-                    for check in checks:
-                        names.append(check.name)                
-                        statuses.append("√ PASSED" if check.success else "X FAILED")
-                        comments.append(check.comment)
-                        ready = ready and check.success
-                    row = [result.connector.name, "\n".join(names), "\n".join(statuses), "\n".join(comments)]
-                    table.append(row)
-            else:                    
-                headers = ["CONNECTOR", "STATUS"]
+                headers = ["CONNECTOR", "CHECK", "ID", "TAGS", "STATUS", "MESSAGE"]                    
                 for result in results:
                     checks: List[Check] = result.value
-                    success = reduce(lambda success, c: success and c.success, checks, True)
-                    ready = ready and success
+                    names, ids, tags, statuses, comments = [], [], [], [], []
+                    for check in checks:
+                        names.append(check.name)
+                        ids.append(check.id)
+                        tags.append(", ".join(check.tags) if check.tags else "-")
+                        statuses.append("√ PASSED" if check.success else "X FAILED")
+                        comments.append(check.message or "-")
+                        ready &= check.success
+                    
+                    if not names:
+                        continue
+
+                    row = [result.connector.name, "\n".join(names), "\n".join(ids), "\n".join(tags), "\n".join(statuses), "\n".join(comments)]
+                    table.append(row)
+            else:                    
+                headers = ["CONNECTOR", "STATUS", "ERRORS"]
+                for result in results:
+                    checks: List[Check] = result.value
+                    if not checks:
+                        continue
+
+                    success = True
+                    errors = []
+                    for check in checks:
+                        success &= check.success
+                        check.success or errors.append(f"{check.name}: {check.message or '-'}")
+                    ready &= success
                     status = "√ PASSED" if success else "X FAILED"
-                    row = [result.connector.name, status]
+                    message = reduce(lambda m, e: m + f"({errors.index(e) + 1}/{len(errors)}) {e}\n", errors, "")
+                    row = [result.connector.name, status, message]
                     table.append(row)
                 
             # Output table and exit