+++ servo/connectors/kubernetes.py
@@ -33,6 +33,8 @@ from servo import (
     Description,
     Duration,
     DurationProgress,
+    Filter,
+    HaltOnFailed,
     License,
     Maturity,
     Setting,
@@ -2478,10 +2480,10 @@ class KubernetesChecks(BaseChecks):
             await KubernetesOptimizations.create(self.config)
         except Exception as e:
             return Check(
-                name="Connect to Kubernetes", success=False, comment=str(e)
+                name="Connect to Kubernetes", success=False, message=str(e)
             )
 
-        return Check(name="Connect to Kubernetes", success=True, comment="")
+        return Check(name="Connect to Kubernetes", success=True, message="")
     
     # TODO: Verify the connectivity & permissions
     # TODO: Check the Deployments exist
@@ -2550,8 +2552,8 @@ class KubernetesConnector(BaseConnector):
             self.logger.info(f"Settlement duration of {settlement} has elapsed, resuming optimization.")
 
     @on_event()
-    async def check(self) -> List[Check]:
-        return await KubernetesChecks.run(self.config)
+    async def check(self, filter_: Optional[Filter], halt_on: HaltOnFailed = HaltOnFailed.requirement) -> List[Check]:
+        return await KubernetesChecks.run(self.config, filter_, halt_on=halt_on)
 
 
 def selector_string(selectors: Mapping[str, str]) -> str: